Make new drink dialog show ingredient list.
